# web_app/backend/openapi.json
{"openapi":"3.1.0","info":{"title":"CAPGuard AI","description":"Production Hybrid AI System for Pediatric Community-Acquired Pneumonia","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/xai/{filename}":{"get":{"summary":"Get Xai Image","operationId":"get_xai_image_api_xai__filename__get","parameters":[{"name":"filename","in":"path","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/health":{"get":{"summary":"Health","operationId":"health_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/models/status":{"get":{"summary":"Models Status","operationId":"models_status_api_models_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/configuration":{"get":{"summary":"Configuration","operationId":"configuration_api_configuration_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/patients":{"get":{"summary":"List Patients","operationId":"list_patients_api_patients_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Create Patient","operationId":"create_patient_api_patients_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/patients/{patient_id}":{"get":{"summary":"Get Patient Endpoint","operationId":"get_patient_endpoint_api_patients__patient_id__get","parameters":[{"name":"patient_id","in":"path","required":true,"schema":{"type":"integer","title":"Patient Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/patients/{patient_id}/xray":{"post":{"summary":"Upload Xray","operationId":"upload_xray_api_patients__patient_id__xray_post","parameters":[{"name":"patient_id","in":"path","required":true,"schema":{"type":"integer","title":"Patient Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_xray_api_patients__patient_id__xray_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/assessment":{"post":{"summary":"Run Assessment","operationId":"run_assessment_api_assessment_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessmentInput"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/history/{patient_id}":{"get":{"summary":"Patient History","operationId":"patient_history_api_history__patient_id__get","parameters":[{"name":"patient_id","in":"path","required":true,"schema":{"type":"integer","title":"Patient Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/history":{"get":{"summary":"All History","operationId":"all_history_api_history_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/assistant/chat":{"post":{"tags":["Clinical AI Assistant"],"summary":"Assistant Chat","operationId":"assistant_chat_api_assistant_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssistantChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssistantChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AssessmentInput":{"properties":{"patient_id":{"type":"integer","title":"Patient Id"},"patient_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Patient Name"},"age":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Age"},"biological_sex":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Biological Sex"},"temperature":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Temperature"},"heart_rate":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Heart Rate"},"oxygen_saturation":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Oxygen Saturation"},"clinical_notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Clinical Notes"},"lab_data":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Lab Data"},"cbc_data":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Cbc Data"},"xray_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Xray Path"}},"type":"object","required":["patient_id"],"title":"AssessmentInput"},"AssistantChatRequest":{"properties":{"question":{"type":"string","maxLength":2000,"minLength":1,"title":"Question"},"language":{"type":"string","pattern":"^(en|ar)$","title":"Language","default":"en"},"patient":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Patient"},"assessment":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Assessment"},"visit":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Visit"}},"type":"object","required":["question"],"title":"AssistantChatRequest"},"AssistantChatResponse":{"properties":{"assistant":{"type":"string","title":"Assistant"},"language":{"type":"string","title":"Language"},"safety_flag":{"type":"boolean","title":"Safety Flag"},"safety_message":{"type":"string","title":"Safety Message"},"context_used":{"additionalProperties":true,"type":"object","title":"Context Used"},"status":{"type":"string","title":"Status"},"treatment_plan":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Treatment Plan"}},"type":"object","required":["assistant","language","safety_flag","safety_message","context_used","status"],"title":"AssistantChatResponse"},"Body_upload_xray_api_patients__patient_id__xray_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_xray_api_patients__patient_id__xray_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PatientCreate":{"properties":{"patient_name":{"type":"string","minLength":1,"title":"Patient Name"},"age":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Age"},"biological_sex":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Biological Sex"}},"type":"object","required":["patient_name"],"title":"PatientCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}